refactor: split backend route monoliths into endpoint modules (winzy.ai-ypya)
Extract social-service (1338→41 lines, 4 endpoint files), habit-service
(1236→41 lines, 5 endpoint files), challenge-service (473→38 lines, 1
endpoint file). Adopt shared primitives: TryGetUserId, TryReadBodyAsync,
JsonDefaults.CamelCase, AddServiceHttpClient. Pure structural move.
